backend/service/requirements.txt
aiofiles
alembic
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
async-property
attrs
Authlib
bcrypt
cached-property
cachetools
certifi
cffi
chardet
charset-normalizer
click
cryptography
deprecation
dnspython
docstring_parser
ecdsa
email_validator
et_xmlfile
Faker
fastapi
fastapi-cli
FastAPI-SQLAlchemy
google-auth
greenlet
gunicorn
h11
httpcore
httptools
httpx
idna
importlib_metadata
iniconfig
itsdangerous
Jinja2
jwcrypto
Mako
markdown-it-py
MarkupSafe
mdurl
minio
mysql-connector-python
numpy
openpyxl
orjson
packaging
pandas
passlib
pluggy
psutil
psycopg2-binary
py
pyasn1
pyasn1_modules
pycparser
pycryptodome
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
Pygments
PyJWT
pymongo
pyodbc
pyparsing
pytest
python-dateutil
python-dotenv
python-editor
python-keycloak
python-multipart
pytz
PyYAML
requests
requests-toolbelt
rich
rich-toolkit
rsa
setuptools
shellingham
six
sniffio
SQLAlchemy
starlette
text-unidecode
toml
typer
typing-inspection
typing_extensions
tzdata
ujson
urllib3
uvicorn
uvloop
watchfiles
websockets
wheel
zipp
langchain
langchain-ollama
torch
transformers
openai